chore: move suppression rules
