+++ frontend/src/scenes/error-tracking/configuration/rules/ErrorTrackingAutoAssignment.tsx
@@ -6,31 +6,37 @@ import { ErrorTrackingAssignmentRule, ErrorTrackingRuleType } from './types'
 
 export function ErrorTrackingAutoAssignment(): JSX.Element {
     return (
-        <ErrorTrackingRules<ErrorTrackingAssignmentRule> ruleType={ErrorTrackingRuleType.Assignment}>
-            {({ rule, editable }) => {
-                return (
-                    <LemonCard key={rule.id} hoverEffect={false} className="flex flex-col p-0">
-                        <div className="flex gap-2 justify-between px-2 py-3">
-                            <div className="flex gap-1 items-center">
-                                <div>Assign to</div>
-                                <ErrorTrackingRules.Assignee rule={rule} editable={editable} />
-                                <div>when</div>
-                                <ErrorTrackingRules.Operator rule={rule} editable={editable} />
-                                <div>filters match</div>
+        <>
+            <p>
+                Automatically assign newly created issues based on properties of the exception event the first time it
+                was seen.
+            </p>
+            <ErrorTrackingRules<ErrorTrackingAssignmentRule> ruleType={ErrorTrackingRuleType.Assignment}>
+                {({ rule, editable }) => {
+                    return (
+                        <LemonCard key={rule.id} hoverEffect={false} className="flex flex-col p-0">
+                            <div className="flex gap-2 justify-between px-2 py-3">
+                                <div className="flex gap-1 items-center">
+                                    <div>Assign to</div>
+                                    <ErrorTrackingRules.Assignee rule={rule} editable={editable} />
+                                    <div>when</div>
+                                    <ErrorTrackingRules.Operator rule={rule} editable={editable} />
+                                    <div>filters match</div>
+                                </div>
+                                <ErrorTrackingRules.Actions rule={rule} editable={editable} />
                             </div>
-                            <ErrorTrackingRules.Actions rule={rule} editable={editable} />
-                        </div>
-                        <LemonDivider className="my-0" />
-                        <div className="p-2">
-                            <ErrorTrackingRules.Filters
-                                taxonomicGroupTypes={[TaxonomicFilterGroupType.EventProperties]}
-                                rule={rule}
-                                editable={editable}
-                            />
-                        </div>
-                    </LemonCard>
-                )
-            }}
-        </ErrorTrackingRules>
+                            <LemonDivider className="my-0" />
+                            <div className="p-2">
+                                <ErrorTrackingRules.Filters
+                                    taxonomicGroupTypes={[TaxonomicFilterGroupType.EventProperties]}
+                                    rule={rule}
+                                    editable={editable}
+                                />
+                            </div>
+                        </LemonCard>
+                    )
+                }}
+            </ErrorTrackingRules>
+        </>
     )
 }
